Removed UInt8 conformance to LosslessStringConvertible
